requirements.txt
streamlit==1.25.0
fastapi==0.103.1
uvicorn==0.22.0
PyPDF2==3.0.1
faiss-cpu==1.7.3
sentence-transformers==2.2.2
openai==0.27.8
